refact: move :app tests to com.jeanbarrossilva.ongoing.app
