+++ app/src/androidTest/java/com/jeanbarrossilva/ongoing/app/extensions/ActivityScenarioExtensions.kt
@@ -1,4 +1,4 @@
-package com.jeanbarrossilva.ongoing.extensions
+package com.jeanbarrossilva.ongoing.app.extensions
 
 import android.app.Activity
 import androidx.test.core.app.ActivityScenario
