+++ app/src/androidTest/java/com/jeanbarrossilva/ongoing/app/extensions/AndroidComposeTestRuleExtensions.kt
@@ -1,4 +1,4 @@
-package com.jeanbarrossilva.ongoing.extensions
+package com.jeanbarrossilva.ongoing.app.extensions
 
 import android.content.Context
 import androidx.activity.ComponentActivity
